# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "XR50TrainingAssetRepo",
    "version": "1.0"
  },
  "paths": {
...
    "/xr50/TrainingAssetRepository/material_management/{MaterialId}": {
      "get": {
        "tags": [
          "material_management"
        ],
        "parameters": [
          {
            "name": "MaterialId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Material"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Material"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Material"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "material_management"
        ],
        "parameters": [
          {
            "name": "MaterialId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/xr50/TrainingAssetRepository/material_management/workflow/{tenantName}": {
      "get": {
        "tags": [
          "material_management"
        ],
        "parameters": [
          {
            "name": "tenantName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkflowMaterial"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkflowMaterial"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkflowMaterial"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/xr50/TrainingAssetRepository/material_management/workflow/{tenantName}/{materialId}": {
      "get": {
        "tags": [
          "material_management"
        ],
        "parameters": [
          {
            "name": "tenantName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "materialId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowMaterial"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowMaterial"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowMaterial"
                }
              }
            }
          }
        }
      }
    },
    "/xr50/TrainingAssetRepository/material_management/{TenantName}": {
      "post": {
        "tags": [
          "material_management"
        ],
        "parameters": [
          {
            "name": "TenantName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Material"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Material"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Material"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Material"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Material"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Material"
                }
              }
            }
          }
        }
      }
    },
    "/xr50/TrainingAssetRepository/material_management/{TenantName}/{ParentMaterialId}": {
      "post": {
        "tags": [
          "material_management"
        ],
        "parameters": [
          {
            "name": "TenantName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ParentMaterialId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Material"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Material"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Material"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Material"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Material"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Material"
                }
              }
            }
          }
        }
      }
    },
    "/xr50/TrainingAssetRepository/material_management/{TenantName}/workflow": {
      "post": {
        "tags": [
          "material_management"
        ],
        "parameters": [
          {
            "name": "TenantName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowMaterial"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowMaterial"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowMaterial"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Material"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Material"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Material"
                }
              }
            }
          }
        }
      }
    },
    "/xr50/TrainingAssetRepository/material_management/{TenantName}/checklist": {
      "post": {
        "tags": [
          "material_management"
        ],
        "parameters": [
          {
            "name": "TenantName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChecklistMaterial"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ChecklistMaterial"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ChecklistMaterial"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Material"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Material"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Material"
                }
              }
            }
          }
        }
      }
    },
    "/xr50/TrainingAssetRepository/material_management/{TenantName}/image": {
      "post": {
        "tags": [
          "material_management"
        ],
        "parameters": [
          {
            "name": "TenantName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageMaterial"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageMaterial"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ImageMaterial"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Material"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Material"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Material"
                }
              }
            }
          }
        }
      }
    },
    "/xr50/TrainingAssetRepository/material_management/{TenantName}/video": {
      "post": {
        "tags": [
          "material_management"
        ],
        "parameters": [
          {
            "name": "TenantName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VideoMaterial"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/VideoMaterial"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/VideoMaterial"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Material"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Material"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Material"
                }
              }
            }
          }
        }
      }
    },
...
    "/xr50/TrainingAssetRepository/tenant_management/material-management/{TenantName}/{ParentMaterialId}": {
      "post": {
        "tags": [
          "tenant_management"
        ],
        "parameters": [
          {
            "name": "TenantName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ParentMaterialId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Material"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Material"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Material"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Material"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Material"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Material"
                }
              }
            }
          }
        }
      }
    },
...
    "/xr50/TrainingAssetRepository/tenant_management/material-management/{TenantName}/{ProgramName }/{MateriaId}": {
      "delete": {
        "tags": [
          "tenant_management"
        ],
        "parameters": [
          {
            "name": "TenantName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ProgramName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "MaterialId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ProgramName ",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "MateriaId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChecklistMaterial": {
        "required": [
          "materialType"
        ],
        "type": "object",
        "properties": {
          "tenantName": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "materialName": {
            "type": "string",
            "nullable": true
          },
          "parentId": {
            "type": "string",
            "nullable": true
          },
          "materialList": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "trainingProgramList": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "materialId": {
            "type": "string",
            "nullable": true
          },
          "materialType": {
            "$ref": "#/components/schemas/MaterialType"
          },
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChecklistEntry"
            },
            "nullable": true
          },
          "entryId": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Material": {
        "required": [
          "materialType"
        ],
        "type": "object",
        "properties": {
          "tenantName": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "materialName": {
            "type": "string",
            "nullable": true
          },
          "parentId": {
            "type": "string",
            "nullable": true
          },
          "materialList": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "trainingProgramList": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "materialId": {
            "type": "string",
            "nullable": true
          },
          "materialType": {
            "$ref": "#/components/schemas/MaterialType"
          }
        },
        "additionalProperties": false
      },
      "MaterialType": {
        "enum": [
          0,
          1,
          2,
          3,
          4
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "WorkflowMaterial": {
        "required": [
          "materialType"
        ],
        "type": "object",
        "properties": {
          "tenantName": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "materialName": {
            "type": "string",
            "nullable": true
          },
          "parentId": {
            "type": "string",
            "nullable": true
          },
          "materialList": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "trainingProgramList": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "materialId": {
            "type": "string",
            "nullable": true
          },
          "materialType": {
            "$ref": "#/components/schemas/MaterialType"
          },
          "steps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkflowStep"
            },
            "nullable": true
          },
          "stepIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}